refactor: :recycle: group and autocomplete updates
