+++ locales/en/filters.json
@@ -44,6 +44,7 @@
     "email": "email",
     "group_role_title": "Usergroup Roles",
     "institution_title": "Institution",
+    "name_of_user": "Name of User",
     "phone": "Phone Number",
     "profession_title": "Profession",
     "role_title": "Roles",
