+++ locales/hi/filters.json
@@ -44,6 +44,7 @@
     "email": "ईमेल",
     "group_role_title": "उपयोगकर्ता समूह भूमिकाएँ",
     "institution_title": "संस्थान",
+    "name_of_user": "उपयोगकर्ता का नाम",
     "phone": "फ़ोन नंबर",
     "profession_title": "पेशा",
     "role_title": "भूमिकाएँ",
