+++ locales/hi/group.json
@@ -5,12 +5,12 @@
     "title": "तकरीबन"
   },
   "admin": {
-    "error": "प्रशासकों को अद्यतन करते समय कुछ त्रुटि हुई",
+    "error": "उपयोगकर्ताओं को अपडेट करते समय कुछ त्रुटि हुई",
     "founder": "संस्थापकों",
     "moderator": "मध्यस्थ",
     "no_access": "व्यवस्थापक से संपर्क करें",
-    "title": "समूह प्रशासक",
-    "updated": "समूह प्रशासकों ने सफलतापूर्वक अद्यतन किया",
+    "title": "समूह उपयोगकर्ता",
+    "updated": "समूह उपयोगकर्ता सफलतापूर्वक अपडेट किए गए",
     "view_members": "सभी सदस्यों को देखें"
   },
   "confirm_closed_group": "यह एक बंद समूह है, सदस्यता अनुरोध को मॉडरेटर द्वारा अनुमोदित किया जाना है।अनुरोध में शामिल होने के लिए ओके दबाएं",
@@ -98,5 +98,5 @@
   "spatial_coverge": "स्थानिक कवरेज",
   "tags": "टैग",
   "update": "अद्यतन समूह",
-  "update_admin": "व्यवस्थापक अद्यतन करें"
+  "update_admin": "उपयोगकर्ताओं को अद्यतन करें"
 }
